backend/requirements.txt
asgiref==3.11.0
charset-normalizer==3.4.4
Django==4.2.27
django-cors-headers==4.9.0
djangorestframework==3.16.1
numpy==2.0.2
pandas==2.3.3
pillow==11.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.4.9
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
whitenoise
gunicorn
matplotlib